Another leetcode problem solved
Do not pay attention to the % that shows how much people you beat because it is a lie. When you submit your work it always runs the program with different speed and memory use, and also the % is also different. Just try to complete the task and maybe complete it again in the hard way.